Fix: handle verbose include/exclude patterns in pyproject.toml
If a verbose pattern was being used in pyproject.toml, the
patterns did not function correctly. For instance, the pattern

exclude = '''
(
/(
  folder1
| folder2
)/
)
'''

does not function properly. This pattern is inspired by the
pattern that black themselves use:
https://github.com/python/black/blob/master/pyproject.toml

Solve this issue by removing newlines, and formatting
the pattern before using it. This is done the same way
as black themselves are doing it.